pyproject.toml
[project]
name = "persona-agent"
version = "0.2.0"
description = "AI persona agents built on Google A2A protocol with FastAPI REST API and MCP tool integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # A2A protocol
    "a2a-sdk>=0.3.25",
    # LLM client
    "openai>=1.30.0",
    # MCP tool integration
    "mcp>=1.0.0",
    # API framework
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "sse-starlette>=2.0.0",
    # Data processing
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    # Utilities
    "httpx>=0.27.0",
    "typing-extensions>=4.8.0",
]

[project.scripts]
persona-agent = "persona_agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/persona_agent"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["persona_agent"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
